requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
arpa==0.1.0b4
asgiref==3.7.2
asttokens==2.0.5
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
Django==5.0.2
djangorestframework==3.14.0
einops==0.7.0
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.9.0
fonttools==4.43.1
fsspec==2023.4.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
gradio==4.7.1
gradio_client==0.7.0
grpcio==1.59.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
idna==3.4
imageio==2.32.0
importlib-metadata==6.0.0
importlib-resources==6.1.0
ipykernel==6.25.0
ipython==8.12.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.9.0
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
Morfessor==2.0.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.1
pandarallel==1.6.5
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
phoenix-datasets==0.0.1.dev0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
prompt-toolkit==3.0.36
protobuf==4.24.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.31.0
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.2.0
starlette==0.27.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
tokenizers==0.14.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
torchdata==0.7.1
torchmetrics==1.2.0
torchtext==0.16.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.7.1
triton==2.1.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.24.0.post1
wcwidth==0.2.5
websockets==11.0.3
Werkzeug==3.0.1
xmltodict==0.13.0
zipp==3.11.0
